feat: 优化活动报名状态流程及管理员留言回复功能
- 报名状态完整展示：PENDING显示"审核中"、REJECTED显示"联系管理员"、CANCELLED允许重新报名
- 管理端新增重新审核功能，可将已拒绝的报名恢复为待审核状态
- 新增活动留言板功能：用户可联系管理员，管理员可在报名管理中回复留言
- 用户留言及管理员回复在活动详情页展示
- 修复首页Hero区域浅色背景下文字不清晰的问题
- 修复论坛页面"最新评论在详情里查看"标签对比度不足问题
- 修复活动列表和详情页状态标签对比度问题
- 修复管理端报名管理操作按钮换行问题

+++ bickdemo-backend/src/main/java/com/example/bickdemo/dto/TicketFeedbackRequest.java
@@ -12,10 +12,10 @@ import lombok.Data;
 @Data
 public class TicketFeedbackRequest {
 
-    /** 评分（1-5星） */
+    /** 评分（0.5-5星，支持半星） */
     @Min(value = 1, message = "评分最小为 1")
     @Max(value = 5, message = "评分最大为 5")
-    private Integer rating;
+    private Double rating;
 
     /** 反馈内容 */
     @Size(max = 1000, message = "反馈内容不能超过 1000 字符")
